+++ CHANGELOG.md
@@ -31,3 +31,4 @@
 ## Miscellaneous
 
 * Renamed MacOS build from "`Woof-<version>-uni.zip`" to "`Woof-<version>-MacOS-universal.zip`"
+* Savegame description and snapshot are now saved outside the compressed keyframe to speed up populating the Load/Save Game menu pages